Performance improvements and design changes
